Remove snippetsEmu
Snippets emu has problems with indentation, and I don't want to fix each
filetype. I'm going to replace it with snipmate -- what looks to be a
superior implementation of the same idea.

snipMate's snippet files are cleaner, in their own folder, and appear to
avoid comments (possible to prevent problems with formatoptions). It's
also prettier. Otherwise the functionality seems to be about the same.