Add script for controlling monitors
I want to use only the external monitor when it's connected, but my
Dell's monitor button (F1) doesn't give me that option (if I do that, it
switches back to laptop display when I hit a key.

I don't want to have to enter Monitors, so this script is my answer.