Improve tag building for android projects
now we get all of the dependencies as well.

If you pass "android" as the language to continuous_build, it will use
the special android tag building that reads the .project file to get
dependencies and build tags for them. However, the file structure
expected isn't flexible: You need to call from the directory that
contains all of the android projects (so they must all be at the same
level in one folder).

I haven't spent any time figuring out how easy it is to support other
configurations.