Add FollowSymlink command: switch to symlink target
fugitive is only enabled when the current file is in a git repo. Since
some of my config files are links to repo files, when fugitive won't
Gdiff. When fugitive opens a diff from Gstatus, it will use the file in
an existing buffer if possible. If that's the symlink, then the Gdiff
fails.

The FollowSymlink command will wipeout the current buffer so that it
won't be reloaded, and then loads the symlink's target. It tries to use
bufkill's BW to prevent the closing of the window, but if bufkill isn't
installed, it just uses bwipeout.