Add AsyncPython: execute the current file as python
Runs the current file as a python script (relying on the path to find
the python executable).

Uses a python errorformat to pull out relevant line information, but it
only includes the top-level of the exception.

Has to be in after directory so it's loaded after AsyncCommand (which is
a bundle).