Remove the squelching line from Python efm
The last line: \%-G%.%# is meant to suppress some late error messages
that I found could occur e.g. with wxPython and that prevent one from
using :clast to go to the relevant file and line of the traceback.

This isn't a problem for me right now and it also hides all other
output.