Add new plugin: magicmakefile
Automatically uses a relevant makefile. If there isn't a make file in
the current directory, it uses a default one that allows you to build
the current file as a standalone executable.

Sets the makeprg variable to use the autodetected makefile.

Note that this won't walk up the file hierarchy to find a makefile.