Remove invalid diff mappings
Checking &diff only checks if the current buffer is using diff mode.
I need to add an autocmd or ftplugin to get these to be set in diff
mode. I don't know how to do that right now and I don't really want
these, so I'll leave it.

Maybe I could remove the &diff check and use a map like this:

   xnoremap < :if &diff | :diffget | normal < | endif<CR>