Setup yankring config
It clobbers Y by default.

It also places its history file in home but it's not hidden! So make a
hidden folder to put it in, like undo. I couldn't use
yankring_history_file to add a dot prefix and having a folder might be
helpful if I want to disable yankring_share_between_instances.