Use gvim for git in cygwin
Setup the GIT_EDITOR variable so I don't need to set an editor in my
gitconfig.

Alternatively, I could do this:
	git config --system core.editor "gvim --nofork"

Since I'm the only one on the system, that should work.

I do not have cygwin vim installed and Windows vim does not work
with cygwin (it just sits there in the background, blocking).

Use gvim instead (so it launches the gui). Be explicit about forking
(to ensure we can use it for operations that wait for the editor
to return).