Improve P4CheckoutFilesInGit* functions
Use perforce.vim if available. This opens another buffer with the
checkout message so you can see if it succeeded.

Be consistent in how we edit - extract into a function. This isolates
the filename escaping (note that this doesn't break git diffs even
though perforce depot paths aren't file paths) and the method we shell
out to perforce.