Use <Leader>e for BufExplorer
vim-vinegar lets me use - to launch netrw and it's consistent with
netrw. That makes <Leader>e for netrw redundant. Use it for BufExplorer
instead (especially since changing my leader makes it harder to hit:
<Space>be is much harder to hit than \be).

It would be neat to split for netrw (and use VinegarUp if possible for
my split window), but that's used for golden-ratio which I use more. I
don't know if I've ever used that map in its previous form.