Replace punctuation textobjs with textobj-between
The pairs of identical character maps are superseded by
vim-textobj-between: im and am. A dedicated textobj is more consistent
(works for all characters) and this functionality is coming in a future
version of vim.

Use the same map for textobj-between that vim_dev is using for the new
built-in textobj: m for matched.