Make cpp SwitchSourceHeader fall back to ctrlp
I'm using a kien's implementation for switching to header files. I
modified it to not auto-accept (since that was always opening the first
file in my filelist -- possibly because of ctrlp_lazy_update), use the
default directory (I want it to work project-wide), and only match the
base filename (no file extension needed since ctrlp doesn't show the
current file).

I changed SwitchSourceHeader to fallback to CtrlPSameName if no header
is found.