Define VISUAL in cygwin
It's used by bash for C-x C-e (edit current line in editor). I guess it
only makes sense for a visual editor so it uses VISUAL? Oh remnants of
the past...

I'm following the same format as unix.bashrc where VISUAL is defined and
everything takes that value. Seems like it's the more important one...