Add whole word matching for CtrlPSameName
Files with names that are common substrings are hard to find with
CtrlPSameName. Add an N variant that searches for whole words.

Update A-o related uses to require whole word match. That's probably
what the file's called. If not, I can use C-o n to non whole word match
(which is easier to type than C-o N).