Automatically run :compiler based on filetype
Previously, I would do this individually with multiple ftplugins, but
this autocmd lets me set it and forget it (unless I need a compiler with
a name that doesn't match the filetype -- like C++). Those cases usually
require thinking about which compiler I want to select.

The downside is that it's harder to use make, scons, or another build
system since this code will change 'makeprg' to the interpreter.
:ForceSetCompiler lets you override the automatic behaviour and always
use a specific compiler for all files.

The silent! is necessary to ignore filetypes like vim or cpp that don't
have a compiler named after the filetype, but will work for many other
language and save me some copypaste for new filetypes.

It overrides any compilers setup in ~/.vim/ftplugin, but not
~/.vim/after/ftplugin which is where I usually invoke :compiler.

I ran across this wikia page[1] that does it with BufRead and
try..catch, but it seems to make more sense to use FileType to me (so it
also works if I change to a different filetype and I know the filetype
has been detected).

[1]: http://vim.wikia.com/wiki/Autoselect_the_right_compiler_using_the_filetype