Add general vim thoughts
Pulled a lot of content from reddit posts, but this is a decent
collection of vim knowledge. There's an unfortunate amount of plugin
links that make it not really good fodder for a newbie, but it's
essentially a brain dump.

Not sure if this should be a wiki or a blog series.