Only cygpath explicitly relative paths
This python script converts unix paths to windows paths for use with
windows python. But if you have an argument that matches a file in the
working directory (ignoring file extensions for executable types!), then
it will use the absolute path to that file instead.

Only cygpath absolute file paths or explicitly relative paths (with ./).

Also echo the command so I don't pull my hair out if something like this
happens again.