Remove nocompatible
I've had nocompatible for a long time and since it was at the beginning
of my vimrc I thought it would cause no trouble. I figured it was useful
if I sourced my vimrc via `vim -u vimrc`. But whenever I source a vimrc,
I always pass -N (for nocompatible) and I've never sourced this vimrc.

itchyny pointed out (in response to someone else's config) that setting
nocompatible in your vimrc, means that :source ~/.vimrc will truncates
the command history to Vim's default value:

http://www.reddit.com/r/vim/comments/2z2fg0/need_some_inspiration_for_your_vimrc/cpf8zho

I've occasionally seen commands that I expect to be in my history to be
not in my history and this is likely the cause. Since it's not giving me
any benefit, it's gone.