Maximize the window and equalize the windows
I prefer to merge with a large frame with each window being of equal
size. simalt only functions on Win32, so include lines/columns for unix.
If you use vim, then only the wincmd = (to make windows the same size)
should do anything (since terminal vim is bound to the size of
the terminal).