Add 'bin' diff type to use hexdump
hexdump is on cygwin and likely on most unixes, so add that to make it
easier to diff binaries. Using an external diff is probably preferable,
but this seems useful. (But the diff is really slow.)

Has no effect unless gitattributes is set with this diff mode enabled.
Ideally, I'd make this more optional (prompt before conversion), but I
don't yet know how.

Source:
http://superuser.com/questions/706042/how-can-i-diff-binary-files-in-git