Change input order; add diffconflicts to vimmerge
Change input order to be more sensible instead of following beyond
compare's crazy: make the local and remote files first (as they are on
diffs) and then the additional files after.

Also setup to use diffconflicts which is a nice way to merge that
includes the automatic merging. It doesn't work great for perforce (the
conflict marker style is different), but it's not too bad.