Don't mess with cwd paths on vim startup
autochdir during vim startup was breaking diffconflicts[1].

The two files it passes are in the same folder. Vim loads the first
fine, but the second one is loaded from cwd (~ or perforce install
directory) instead of its containing folder -- so the file was not
found.

Not sure if this is just a bug from my current version of vim
(7.4.1832) or a knockon from a recent change because I don't see how my
dirvish changes could have broken this: I think I was only *added*
restrictions and didn't remove any.

 [1]: multi/git/tool/mergetool.diffconflicts.git.sh