Try using s instead of c for surround
Try using s instead of c. See how it works. c is more useful in general
(since it's an operator instead of a single command), so I should get
use to using it consistently.

After a few weeks, this has been hard but the hardest part is having
other vimconfigs that aren't setup like this. Push it out and I can
revert later.