Remove Argumentative and all uses
I've swapped from Argumentative to Sideways. Sideways seems to handle
some corner/malformed cases better which was causing me headaches
before.

In theory, Sideways can do more stuff than Argumentative (and it's twice
the amount of code), but I haven't tried it.

Did a search for argumentative in daveconfig and found only comments of
how I'm not using it anymore.