Use recurseSubmodules=check to prevent bad pushes
I use a lot of submodules (especially for daveconfig), so I'd like git
to prevent me from making bad pushes (where submodule remote doesn't
have my changes linked to in gitmodules).

Unfortunately, the setting doesn't seem to work on git 2.4.5 (what I
have on cygwin). The cmdline args work, but the setting does not.

    $ git push --recurse-submodules=check mine master
    The following submodule paths contain changes that can
    not be found on any remote:
      multi/vim/bundle/unite-david

    Please try

            git push --recurse-submodules=on-demand

    or cd to the path and use

            git push

    to push them to a remote.

    fatal: Aborting.
    $ git config --get push.recurseSubmodules
    check
    $ git push mine master
    Counting objects: 5, done.
    Delta compression using up to 8 threads.
    Compressing objects: 100% (5/5), done.
    Writing objects: 100% (5/5), 477 bytes | 0 bytes/s, done.
    Total 5 (delta 4), reused 0 (delta 0)
    remote: Resolving deltas: 100% (4/4), completed with 4 local objects.
    To git@github.com:idbrii/daveconfig.git
       6b8ffe2..b1ec528  master -> master

The
[answer](http://stackoverflow.com/questions/5814319/git-submodule-push)
I was looking at says the option exists in git 2.7.

Updated cygwin and verified that it works in git 2.8.3.