Change buildtags to prune first and match second
After reading through this writeup on prune[1], I think one of the major
things to remember is that you should prune before matching anything
else.

Also, we prune directories, so we should only accept directories for
pruning.

I'm holding off on updating the C# version until I actually have a
project to run it on.

[1]: https://stackoverflow.com/a/1489405/79125