Add partial support for loa hotswap errors
These errors manifest in one of two ways:
    [string "cannot OLDFILEACCESSMETHOD @scripts/screens/mainscreen.lua..."]:234: 'end' expected near 'self'
    error	#[string "scripts/widget/util.lua"]:77: variable 'Text' is not declared
The second is similar to existing assert capture, so simplify that to
chomp more leading text.

The first is a bit special so add a new case for it. However, this
doesn't always work: sometimes we don't get a proper filename:
    [string "cannot OLDFILEACCESSMETHOD @scripts/reload...."]:33: table index is nil

Don't know how to deal with that. For now, we capture the error but jump
to a nonexistent file. That's good enough: user will do better at
figuring out filename.