Guard foldmethod changes against diff in ftplugins
When setting foldmethod in ftplugins, we risk overriding the diff
foldmethod (if it was set before FileType occurred). Guard against this
in all of my ftplugins.

I specifically encountered this problem with vim-vc (svn source control)
on a python file and I previously encountered it with a lua file
(3cc0a028). Not sure if I've ever encountered with fugitive.