sub: Fork pydoc from newest fork and add shellslash
Original maintainer hasn't pushed a change since 2016. New somini has a
change in 2019 and did a bunch of work to change the plugin including
python version detection. Not sure if the changes are all good (it's no
longer an ftplugin?). None of the other forks have anything interesting
that isn't merged into fs111 or somini's repos.

Push my shellslash fix to my repo. I don't know if it makes sense as a
PR yet.