Update mac bashrc to match newer setup
Just use profile to open a bashrc and make it like the other ones.
Doesn't work well (no terminal colours), so I have some troubleshooting
to do.

mac.bashrc is just what used to be in profile.

Apparently this is closer to how bash init is supposed to work? (profile
for system-wide and nonbash things... which isn't much).