Add a new VS theme: SublimeDavid
I forgot about my existing themes, so I made a new one off
SublimeMaterialDark. My only modification is to change the title bar to
show the window focus (dark blue when focused and grey when not) and the
icon shows the build state.

First, add the SublimeMaterialDark original version from
https://raw.githubusercontent.com/Microsoft/VS-ColorThemes/03afcb60f9adff69709971268f5ca16f4527e9f6/VSColorThemes/Themes/SublimeMaterialDark.xml