Prevent loading py2 code from py3
Using both python 2 and 3 made makeprg=python not work: Whichever I
setup first will be the makeprg, but it seems the other will get
imported in python. py2 barfs on py3 scripts (site.py) and py3 barfs on
py2 scripts (encodings/__init__.py).

That was because I was adding py2 to py3's PYTHONPATH, but we don't want
that.

Fixing this seems to prevent using both in vim (unsurprisingly). I could
have both in vim, but not both for :make. Oh well.