Fix ahk golden ratio behavior
I don't know if this was working at all before? It doesn't seem to
recognize my golden_ratio variable at all (at least it didn't appear in
listvars -- maybe because it wasn't marked as global) and width was
always 0.

ratios seemed to be backwards. 1/1.6 ~= 0.6 so that's the big one.

Now it works, but only for the primary monitor.