Provide a / easy-align delimiter for comments
I often run into problems trying to align comments (adding a leading
space is the solution I remember and it makes sense that doing so adds a
character that's not part of the comment). This gives me a dedicated
command to align on comments. Mnemonic is the same as vim-commentary: /
for c-style comments.