Fix giv's argument passing
giv was causing "--not: command not found" because I need to quote
variable assignment.

Not sure how to get args to work as desired since vim needs them to be
quoted but doesn't like the various ways I've tried to quote it, so I'm
using two branches.

Can't pass refs that don't exist or git will fail, so we need to check
whether the log command will pass before attempting it.

Remove the trailing . because it just makes vim report "." is a
directory.