Make a very simple alternative to hub/gh
I don't really want all the features of gh (github/cli). I just want it
to be easier to add remotes and to create forks. This solves the first
problem. Can't solve the second without using the API.

This command is my attempt to expose some functionality without
interacting with github. Not sure the name is great since it doesn't
actually create a fork.