Adjust arguments for inotifywait in buildtags
Fix build on editing a file.
Fix spammy "Setting up watches. warning recursive. blah blah"

Still outputs the failure message, so we could do more processing on
that if necessary. For now we ignore it and rely on proper
configuration.

Could ignore .md files by extending the regex, but whatever.