Disable roman and add SpeedDatingRoman to toggle
I rarely want speeddating's roman numeral support so disable it. Add a
command so I can bring it back.

Not sure how bad this will be on startup time, but the alternative is
hacking speeddating -- the docs explicitly mention this method for
removing roman so I assume tpope doesn't want an option.