sub: Add vim-operator-sort
I recently wanted vip:sort to be repeatable. I couldn't figure out how
to do it with advanced-sorters and I ended up doing it with:
    10,51g/{$/norm jvai:sort

But that's pretty awkward, so add a new operator. Looks like this can
also sort within a line.

Replace my :sort maps with operator-sort and add a normal one.

My <Leader>s is pretty overloaded, so use 'so' for sort.