Make original_makeprg buffer-local
Fix SetEntrypoint sometimes uses the wrong makeprg.

SetEntrypoint could be called for all kinds of different makeprg, so we
can only trust the original makeprg for the current file -- not for the
command itself.

We're trying to avoid double adds to the makeprg. Preserve the original
one that should be set by the ftplugin.