Use git-credential-manager
git-credential-manager seems to be the main recommended helper to use
and the git-credential-manager-core one is the old name for it.

Getting a shim for git-credential-manager.exe requires my PR against
Scoop/Extras, but hopefully it gets merged before I install a new
computer.

I think my 2023-01-17 problems were because I installed the wrong one.