nvim: Switch from goneovim -> neovide
Lots of visual glitches with goneovim (looks like failure to fully
repaint window and it's especially bad when doing a Gdiff where you see
horizontal bars while it fetches the index file to diff). neovide has a
slower startup, but if I turn off animations it seems fine. Not sure it
doesn't have the same glitches.