nvim: Switch quickfix-reflector -> quicker
I keep getting errors when applying quickfix changes in nvim. I suspect
some incompatibility? I've had other problems with quickfix-reflector in
the past where it needs me to refresh the quickfix to make it writeable
again. Try a different plugin to see if the grass is greener.

See also corresponing setup() and config in vim-david.