nvim: Add CopilotChat.nvim
Optional and nvim-only just like vim-copilot. Not sure if this will be
useful, but coworkers say this is one of the more useful parts of
copilot.

There's a PR for vim-copilot that adds a similar chat, but it also does
http requests directly against the copilot api so it doesn't seem like
much value to use a fork.