Pass absolute paths to nvim
When calling `vide readme.md`, it was opening files in nvim's working
directory instead of the cwd of the process sending files to nvim. That
seems dumb, but I'm not sure what else I can do about it.

Tried `file=%~f1 %~f2` but then the server start doesn't work because file=" ".